requirements.txt
Django==3.0.4
Django-crispy-forms
django-filter
django-widget-tweaks
django-smart-selects
requests_oauthlib==1.3.0
pyyaml==5.3.1
python-dateutil==2.8.1
django-import-export